deploy: xALPACA to fantom
