+++ deploy/exec/minifl/config/add-pool.ts
@@ -1,14 +1,17 @@
 import { HardhatRuntimeEnvironment } from "hardhat/types";
 import { DeployFunction } from "hardhat-deploy/types";
 import { ethers, network } from "hardhat";
-import { Timelock__factory } from "../../../../typechain";
+import { MiniFL__factory, Timelock, Timelock__factory } from "../../../../typechain";
 import { getConfig } from "../../../entities/config";
+import { TimelockEntity } from "../../../entities";
+import { FileService, TimelockService } from "../../../services";
 
 interface IAddPool {
   STAKING_TOKEN_ADDRESS: string;
   ALLOC_POINT: number;
   REWARDER_ADDRESS: string;
   IS_DEBT_TOKEN: boolean;
+  WITH_UPDATE: boolean;
 }
 
 const func: DeployFunction = async function (hre: HardhatRuntimeEnvironment) {
@@ -23,41 +26,49 @@ const func: DeployFunction = async function (hre: HardhatRuntimeEnvironment) {
   */
   const POOLS: Array<IAddPool> = [
     {
-      STAKING_TOKEN_ADDRESS: "0x50E5748a2E9C5e05Ba8b95549Bf7B11dA91ddAB7",
+      STAKING_TOKEN_ADDRESS: "0xA3451cfC2f630Fc2DD8230f33Dd925c2063d2737",
       ALLOC_POINT: 0,
-      REWARDER_ADDRESS: "0x0000000000000000000000000000000000000000",
+      REWARDER_ADDRESS: "0x7EEAA96bf1aBaA206615046c0991E678a2b12Da1",
       IS_DEBT_TOKEN: false,
+      WITH_UPDATE: true,
     },
   ];
   const EXACT_ETA = "1639738800";
 
   const config = getConfig();
-
-  const timelock = Timelock__factory.connect(config.Timelock, (await ethers.getSigners())[0]);
+  const [deployer] = await ethers.getSigners();
+  const miniFL = MiniFL__factory.connect(config.MiniFL!.address, deployer);
+  const timelockTransactions: Array<TimelockEntity.Transaction> = [];
+  const isTimelocked = (await miniFL.owner()).toLowerCase() === config.Timelock.toLowerCase();
 
   for (const pool of POOLS) {
-    console.log(`>> Timelock: Add ${pool.STAKING_TOKEN_ADDRESS} with ${pool.ALLOC_POINT} via Timelock`);
-    await timelock.queueTransaction(
-      config.MiniFL!.address,
-      "0",
-      "addPool(uint256,address,address,bool)",
-      ethers.utils.defaultAbiCoder.encode(
-        ["uint256", "address", "address", "bool"],
-        [pool.ALLOC_POINT, pool.STAKING_TOKEN_ADDRESS, pool.REWARDER_ADDRESS, pool.IS_DEBT_TOKEN]
-      ),
-      EXACT_ETA
-    );
-    console.log("generate timelock.executeTransaction:");
-    console.log(
-      `await timelock.executeTransaction('${
-        config.MiniFL!.address
-      }', '0', 'addPool(uint256,address,address,bool)', ethers.utils.defaultAbiCoder.encode(['uint256','address','bool'],[${
-        pool.ALLOC_POINT
-      }, '${pool.STAKING_TOKEN_ADDRESS}', '${pool.REWARDER_ADDRESS}', '${pool.IS_DEBT_TOKEN}']), ${EXACT_ETA})`
-    );
-    console.log("✅ Done");
+    if (isTimelocked) {
+      timelockTransactions.push(
+        await TimelockService.queueTransaction(
+          `>> Timelock: Add ${pool.STAKING_TOKEN_ADDRESS} with ${pool.ALLOC_POINT} via Timelock`,
+          config.MiniFL!.address,
+          "0",
+          "addPool(uint256,address,address,bool,bool)",
+          ["uint256", "address", "address", "bool", "bool"],
+          [pool.ALLOC_POINT, pool.STAKING_TOKEN_ADDRESS, pool.REWARDER_ADDRESS, pool.IS_DEBT_TOKEN, pool.WITH_UPDATE],
+          EXACT_ETA
+        )
+      );
+    } else {
+      console.log(`>> Add ${pool.STAKING_TOKEN_ADDRESS} with ${pool.ALLOC_POINT} point`);
+      await miniFL.addPool(
+        pool.ALLOC_POINT,
+        pool.STAKING_TOKEN_ADDRESS,
+        pool.REWARDER_ADDRESS,
+        pool.IS_DEBT_TOKEN,
+        pool.WITH_UPDATE
+      );
+      console.log(">> Done");
+    }
   }
+
+  if (isTimelocked) FileService.write("add-pool", timelockTransactions);
 };
 
 export default func;
-func.tags = ["TimelockMiniFLAddPool"];
+func.tags = ["MiniFLAddPool"];
