+++ package.json
@@ -229,7 +229,7 @@
     "deploy:fantom_testnet:rewarder1:deploy:rewarder1": "hardhat --network fantom_testnet deploy --no-compile --reset --tags Rewarder1",
     "deploy:fantom_testnet:vault:deploy:minifl-vault": "hardhat --network fantom_testnet deploy --no-compile --reset --tags MiniFL_Vault",
     "deploy:fantom_testnet:two-slope-model:deploy:two-slope-model": "hardhat --network fantom_testnet deploy --no-compile --reset --tags TwoSlopeModel",
-    "deploy:fantom_testnet:minifl:config:add-pool-no-timelock": "hardhat --network fantom_testnet deploy --no-compile --reset --tags MiniFLAddPool",
+    "deploy:fantom_testnet:minifl:config:add-pool": "hardhat --network fantom_testnet deploy --no-compile --reset --tags MiniFLAddPool",
     "deploy:fantom_testnet:minifl:config:set-pool-no-timelock": "hardhat --network fantom_testnet deploy --no-compile --reset --tags MiniFLSetPool",
     "deploy:fantom_testnet:workers:config:set-max-reinvest-bouty-bps-workers03": "hardhat --network fantom_testnet deploy --no-compile --reset --tags SetMaxReinvestBountyBpsWorkers03",
     "deploy:fantom_testnet:workers:config:set-beneficial-vault-config-workers03": "hardhat --network fantom_testnet deploy --no-compile --reset --tags SetBeneficialBuybackFieldsWorkers03",
@@ -259,7 +259,7 @@
     "deploy:fantom_mainnet:oracles:config:add-price-source-chainlink": "hardhat --network fantom_mainnet deploy --no-compile --reset --tags AddSourceChainLinkPriceOracle",
     "deploy:fantom_mainnet:oracles:config:add-primary-source-oracle-medianizer": "hardhat --network fantom_mainnet deploy --no-compile --reset --tags AddSourceOracleMedianizer",
     "deploy:fantom_mainnet:two-slope-model:deploy:two-slope-model": "hardhat --network fantom_mainnet deploy --no-compile --reset --tags TwoSlopeModel",
-    "deploy:fantom_mainnet:minifl:config:add-pool-no-timelock": "hardhat --network fantom_mainnet deploy --no-compile --reset --tags MiniFLAddPool",
+    "deploy:fantom_mainnet:minifl:config:add-pool": "hardhat --network fantom_mainnet deploy --no-compile --reset --tags MiniFLAddPool",
     "deploy:fantom_mainnet:minifl:config:set-pool-no-timelock": "hardhat --network fantom_mainnet deploy --no-compile --reset --tags MiniFLSetPool",
     "deploy:fantom_mainnet:workers:config:set-max-reinvest-bouty-bps-workers03": "hardhat --network fantom_mainnet deploy --no-compile --reset --tags SetMaxReinvestBountyBpsWorkers03",
     "deploy:fantom_mainnet:workers:config:set-beneficial-vault-config-workers03": "hardhat --network fantom_mainnet deploy --no-compile --reset --tags SetBeneficialBuybackFieldsWorkers03",
